==> reports_Paper/2_VizAndCreatingEnvRasters.Rmd <==
---
title: "Correlations and raster stacks of environmental variables"
author: "Juliette Archambeau"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    df_print: kable
    highlight: textmate
    number_sections: yes
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
  pdf_document:
    toc: yes
    toc_depth: '4'
editor_options:
  chunk_output_type: console
bibliography: references.bib
always_allow_html: true
---


<style>
pre {
  overflow-x: auto;
}
pre code {
  word-wrap: normal;
  white-space: pre;
}
</style>

<style type="text/css">
div.main-container {
  max-width: 2000px;
  margin-left: auto;
  margin-right: auto;
}
</style>


```{css, echo=FALSE}
pre {
  max-height: 300px;
  overflow-y: auto;
}

pre[class] {
  max-height: 600px;
}
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(cache=F)
options(width = 300)
library(knitr)      # CRAN v1.26
library(kableExtra) # CRAN v1.1.0
library(dplyr)      # CRAN v1.0.0
library(psych)      # CRAN v1.8.12 (pairs.panels function)
library(tidyverse)  # CRAN v1.3.0
library(raster)     # CRAN v3.3-13
library(stringr)    # CRAN v1.4.0
library(ggbiplot)
```

***

<font size="4">Set of environmental covariates used in the present study:</font>

<span style="color: ForestGreen;">**Burned area**</span>: `BurnedArea`, i.e. the average of the monthly burned area from June 1995 to December 2014 (hectares) extracted from the *GFED4* database at 0.25 degrees resolution (∼28 km resolution) [@giglio2013analysis].

<span style="color: ForestGreen;">**Topographic variable**</span>:`TRI`: topographic ruggedness index (unitless).

We used the topographic data generated by the *NASA’s Shuttle Radar Topography Mission* (SRTM) at 90 m resolution. We then used the *SAGA* v2.3.1 [@conrad2015system] to calculate the topographic ruggedness index (TRI; m) which quantifies the terrain heterogeneity, i.e. differences in elevation between adjacent cells [@riley1999index].

<span style="color: ForestGreen;">**Soil-related variables**</span>, extracted from the *European Soil Database Derived data* at 1 km resolution [@hiederer2013mapping]:

  - `water_top` = the total available water content in the topsoil (0-30 cm; mm)
  - `depth_roots` = depth available to roots (cm).


<span style="color: ForestGreen;">**Climatic variables.**</span>
  
To describe the **current** climates to which the populations are adapted to, we used climatic variables from the *WorldClim* database at 30 arc‐seconds spatial resolution average for the years 1970-2000 [@fick2017worldclim]:

 - `bio5` = max temperature of the warmest month (°C)
 - `bio6`: min temperature of the coldest month (°C)
 - `bio12`: annual precipitation (mm)
 - `bio15`: precipitation Seasonality (coefficient of Variation)

The potential **future** climate was described by two shared socio-economic pathways (the moderately alarming SSP3-7.0 and the strongly alarming SSP5-8.5) averaged over the period 2041-2060 at 2.5 arc-minutes spatial resolution. For each of these SSPs, the predictions of the nine global climate models available in WorldClim were averaged to account for scenarios' uncertainty 

***

<font size="4">This set of covariates was selected because:</font>

  - in previous analyses based on a subset of the genomic data (Infinium assay data), this is the set (among 6 sets tested) that shows the highest number of candidate SNPs. 
  
  - this set has biological sense. In particular, it includes `bio6`, the minimum temperature of the coldest month, and studies are accumulating showing that maritime pine patterns of adaptation to climate are associated with minimum temperatures. 

***

In this document, we first visualize the **correlations among the selected environmental variables**. Second, we create **raster stacks** that will be used for projecting the current and future genomic composition across the landscape. The raster stacks are not stored in the DRYAD repository as they are heavy and can be easily built from the databases cited above.

# Correlations among environmnental variables


  
```{r CorrelationsPanels,fig.height=9,fig.width=12}
# We load the dataset with the extracted environmental variables:
data <- read.csv("data_DRYAD/PopulationPopStructureEnvData.csv")

select.var <- c("bio5",
                "bio6",
                "bio12",
                "bio15",
                "water_top",
                "depth_roots",
                "TRI",
                "BurnedArea")

data %>% 
  dplyr::select(all_of(select.var)) %>% 
  unique() %>% 
  pairs.panels(scale=T,hist.col="palegreen1")
```

```{r PlotCorr,fig.height=7.3,fig.width=7.3}
source("scripts/Functions/corpmat.R")
col <- colorRampPalette(c("#BB4444", "#EE9988", "#FFFFFF", "#77AADD", "#4477AA"))

PlotCors <- function(data,select.var,nbsize){
  cor <- cor(data[,select.var])

# matrix of the p-value of the correlation
p.mat <- corpmat(cor)

corrplot::corrplot(cor, method="color", col=col(200),  
                   type="upper", order="hclust", 
                   addCoef.col = "black", # Add coefficient of correlation
                   tl.col="black", tl.srt=45, #Text label color and rotation
                   # Combine with significance
                   p.mat = p.mat, sig.level = 0.01, insig = "blank", 
                   # hide correlation coefficient on the principal diagonal
                   diag=FALSE,number.cex=nbsize)

pca <- prcomp(data[,select.var], center = TRUE,scale. = TRUE)

ggbiplot(pca,varname.size =4) +  ylim(-2.5, 2.5) +    xlim(-3, 3) +  
    theme_minimal(base_size = 12) +
    theme(plot.title = element_text(size=12))
}

data %>% 
  dplyr::select(all_of(select.var)) %>% 
  unique() %>% 
  PlotCors(select.var = colnames(.),nbsize = 0.6)
```

# Building raster stacks


```{r ClimaticVariables}
# Climatic variables from WordClim
clim.var <- c("bio5","bio6","bio12","bio15")
```

```{r FunctionRemoveNAsStack}
remove.NAs.stack<-function(rast.stack){
  nom<-names(rast.stack)
  test1<-calc(rast.stack, fun=sum)
  test1[!is.na(test1)]<-1
  test2<-rast.stack*test1
  test2<-stack(test2)
  names(test2)<-nom
  return(test2)
}
```


Load the mask of the maritime pine distribution (made with Euforgen data and the NFI plots)

```{r LoadMask}
PinpinDistri <- shapefile('data_DRYAD/Mapping/PinpinDistriEUforgen_NFIplotsBuffer10km.shp')
plot(PinpinDistri)
```


`r knitr::opts_chunk$set(eval = F)`

## Current climate (res 30 seconds)

We create raster stacks with the topographic, soil, fire-related variables and **current** climatic conditions from WorldClim (i.e. from 1970 to 2000).

```{r CurrentClimateRes30sec}

# Climatic rasters
# ----------------
grids <- list.files("data/climate/CurrentClimate/WorldClim30sec_1970_2000/")
grids <- grids[grepl("bio_5|bio_6|bio_12|bio_15",grids)==T]
grids <- c(grids[[3]],grids[[4]],grids[[1]],grids[[2]]) # reorder
rast.clim <- raster::stack(paste0("data/climate/CurrentClimate/WorldClim30sec_1970_2000/", grids))
names(rast.clim) <- clim.var
rast.clim <- crop(rast.clim,extent(PinpinDistri)) 
rast.clim <- mask(rast.clim,PinpinDistri, updatevalue=NA)


# Soil rasters
# ------------
rast.soil <- stack("data/soil/STU_EU_T_TAWC_WGS84.tif","data/soil/STU_EU_DEPTH_ROOTS_WGS84.tif")
names(rast.soil) <- c("water_top","depth_roots")
rast.soil <- raster::crop(rast.soil,extent(rast.clim))   # attribute the extent of climatic raster to soil raster
rast.soil <- raster::resample(rast.soil,rast.clim)       # same number of cells between the soil and climatic rasters
rast.soil <- mask(rast.soil,PinpinDistri,updatevalue=NA) # keep only cells within maritime pine distribution 


# TRI raster
# ----------
# > The models are fitted with TRI values from raster at 90m resolution, and the 
# projections will be on raster at 1km resolution (to have the same resolution as WorldClim variables).
rast.tri <- raster("data/Topography/TRI_WGS84.tif")
rast.tri <- crop(rast.tri,extent(rast.clim))           # attribute the extent of climatic raster to TRI rasters
rast.tri <- raster::resample(rast.tri,rast.clim)       # same number of cells between the TRI and climatic rasters
names(rast.tri) <- c("TRI")
rast.tri <- mask(rast.tri,PinpinDistri,updatevalue=NA) # keep only cells within maritime pine distribution 


# Burned Area raster
# ------------------
rast.BA <- raster("data/FireData/GFED4/PreparedRasters/GFED4_199507_201412_AllCountries.grd")
rast.BA <- crop(rast.BA,extent(rast.clim))
rast.BA <- raster::resample(rast.BA,rast.clim)
names(rast.BA) <- "BurnedArea"
rast.BA <- mask(rast.BA,PinpinDistri,updatevalue=NA)


# Merge and save the rasters in a stack (not included in the DRYAD repository)
# ----------------------------------------------------------------------------
stackall <- stack(rast.clim,rast.soil,rast.tri,rast.BA) # merge in a stack the different layers
stackall <- remove.NAs.stack(stackall) # same cells with NAs in all layers
writeRaster(stackall, filename="data/StacksEnvVars/Stack_CurrentClimate_Res30seconds.grd", format="raster",overwrite=TRUE)
```


## Future climate (res 2.5 minutes)

> Period: 2041-2060

Predictions of **future** climate come from the nine global climate models (GCMs) available in WorldClim for the period 2041 to 2060 and the Shared-socio Economic Pathways (SSP) 370 and 585 (https://www.worldclim.org/data/cmip6/cmip6climate.html).

For each global climate model, we create a **stack of 8 environmental variables**: the 4 variables describing the future climate (`bio5`, `bio6`, `bio12` and `bio15`), the soil-related variables (`water_top` and `depth_roots`), the topographic rugedness index (`TRI`) and the burned area (`BurnedArea`).

### SSP3-7.0 

```{r FutureClimateSSP370}
rast.to.crop <- raster("data/StacksEnvVars/Stack_CurrentClimate_Res30seconds.grd")

# The nine global climate models:
GCMs <- list.files("data/climate/FutureClimate/share/spatial03/worldclim/cmip6/7_fut/2.5m/")
GCMs

list.rast.GCMs <- sapply(GCMs, function(i){
  
  rast.clim <- stack(paste0("data/climate/FutureClimate/share/spatial03/worldclim/cmip6/7_fut/2.5m/",i,"/ssp370/wc2.1_2.5m_bioc_",i,"_ssp370_2041-2060.tif"))
  
  # I understand that these 19 variables here are the WorldClim bioclimatic variables averaged over the period 2041/2060.
  
  select.layers <- names(rast.clim)[grepl("2060.5$|2060.6|2060.12|2060.15",names(rast.clim))==T]
  rast.clim <- raster::subset(rast.clim,select.layers)
  names(rast.clim) <- clim.var
  #rast.clim <- rast.clim[[clim.var]]
  rast.clim <- raster::crop(rast.clim,extent(rast.to.crop))
  rast.clim <- mask(rast.clim,PinpinDistri, updatevalue=NA)
}, USE.NAMES = TRUE,simplify=FALSE)
  

# Soil rasters
rast.soil <- stack("data/soil/STU_EU_T_TAWC_WGS84.tif","data/soil/STU_EU_DEPTH_ROOTS_WGS84.tif")
names(rast.soil) <- c("water_top","depth_roots")
rast.soil <- raster::crop(rast.soil,extent(list.rast.GCMs[[1]]))
rast.soil <- raster::resample(rast.soil,list.rast.GCMs[[1]])
rast.soil <- mask(rast.soil,PinpinDistri,updatevalue=NA)
  
# TRI raster
rast.tri <- raster("data/Topography/TRI_WGS84.tif")
rast.tri <- crop(rast.tri,extent(list.rast.GCMs[[1]]))
rast.tri <- raster::resample(rast.tri,list.rast.GCMs[[1]])
names(rast.tri) <- c("TRI")
rast.tri <- mask(rast.tri,PinpinDistri,updatevalue=NA)
  
# Burned Area
rast.BA <- raster("data/FireData/GFED4/PreparedRasters/GFED4_199507_201412_AllCountries.grd")
rast.BA <- crop(rast.BA,extent(list.rast.GCMs[[1]]))
rast.BA <- raster::resample(rast.BA,list.rast.GCMs[[1]])
names(rast.BA) <- "BurnedArea"
rast.BA <- mask(rast.BA,PinpinDistri,updatevalue=NA)
  

for(i in GCMs){
  
  stackall <- stack(list.rast.GCMs[[i]],rast.soil,rast.tri,rast.BA)
  stackall <- remove.NAs.stack(stackall) # same cells with NAs in all layers
  
  writeRaster(stackall, 
              filename=paste0("data/StacksEnvVars/FutureClimate/2041to2060/SSP370/Stack_",str_remove_all(i,"-"),".grd"),
              format="raster",
              overwrite=TRUE)
  
}
```


### SSP5-8.5

```{r FutureClimateSSP585}
# We remove the GCM that is not included in this SSP:
GCMs <- GCMs[!GCMs=="GFDL-ESM4"]

list.rast.GCMs <- sapply(GCMs, function(i){
  
  rast.clim <- stack(paste0("data/climate/FutureClimate/share/spatial03/worldclim/cmip6/7_fut/2.5m/",i,"/ssp585/wc2.1_2.5m_bioc_",i,"_ssp585_2041-2060.tif"))
  
  # I understand that these 19 variables here are the WorldClim bioclimatic variables averaged over the period 2041/2060.
  
  select.layers <- names(rast.clim)[grepl("2060.5$|2060.6|2060.12|2060.15",names(rast.clim))==T]
  rast.clim <- raster::subset(rast.clim,select.layers)
  names(rast.clim) <- clim.var
  #rast.clim <- rast.clim[[clim.var]]
  rast.clim <- raster::crop(rast.clim,extent(rast.to.crop))
  rast.clim <- mask(rast.clim,PinpinDistri, updatevalue=NA)
}, USE.NAMES = TRUE,simplify=FALSE)
  

# Soil rasters
rast.soil <- stack("data/soil/STU_EU_T_TAWC_WGS84.tif","data/soil/STU_EU_DEPTH_ROOTS_WGS84.tif")
names(rast.soil) <- c("water_top","depth_roots")
rast.soil <- raster::crop(rast.soil,extent(list.rast.GCMs[[1]]))
rast.soil <- raster::resample(rast.soil,list.rast.GCMs[[1]])
rast.soil <- mask(rast.soil,PinpinDistri,updatevalue=NA)
  
# TRI raster
rast.tri <- raster("data/Topography/TRI_WGS84.tif")
rast.tri <- crop(rast.tri,extent(list.rast.GCMs[[1]]))
rast.tri <- raster::resample(rast.tri,list.rast.GCMs[[1]])
names(rast.tri) <- c("TRI")
rast.tri <- mask(rast.tri,PinpinDistri,updatevalue=NA)
  
# Burned Area
rast.BA <- raster("data/FireData/GFED4/PreparedRasters/GFED4_199507_201412_AllCountries.grd")
rast.BA <- crop(rast.BA,extent(list.rast.GCMs[[1]]))
rast.BA <- raster::resample(rast.BA,list.rast.GCMs[[1]])
names(rast.BA) <- "BurnedArea"
rast.BA <- mask(rast.BA,PinpinDistri,updatevalue=NA)
  

for(i in GCMs){
  
  stackall <- stack(list.rast.GCMs[[i]],rast.soil,rast.tri,rast.BA)
  stackall <- remove.NAs.stack(stackall) # same cells with NAs in all layers
  
  writeRaster(stackall, 
              filename=paste0("data/StacksEnvVars/FutureClimate/2041to2060/SSP585/Stack_",str_remove_all(i,"-"),".grd"), format="raster",overwrite=TRUE)
  
}
```

 
### Averaging GCMs predictions

We average the predictions of the nine (SSP3-7.0) or eight (SSP5-8.5) GCMs.


```{r FunctionAveragingGCMs}
# Function to average the predictions of the nine or eight GCMs

AverageGCMs <- function(SSP){

  path=paste0("data/StacksEnvVars/FutureClimate/2041to2060/",SSP,"/")
  file.names <- list.files(path=path,pattern="*grd$")

                    # Create a stack with nine layers (each corresponding to one GCM),
  # and in each layer, there is a stack with the 8 environmental variables
  stack.GCMs <- lapply(file.names, function(t) stack(paste0(path,t)))

  meanIgnoringZeroes <- function(x) {mean(x[x!=0],na.rm=T)}

  if(SSP=="SSP370"){
  AveragingClimGCMs <- function(clim.var,clim.stack){
    mean.clim <- overlay(
          clim.stack[[1]][[clim.var]],
          clim.stack[[2]][[clim.var]],
          clim.stack[[3]][[clim.var]],
          clim.stack[[4]][[clim.var]],
          clim.stack[[5]][[clim.var]],
          clim.stack[[6]][[clim.var]],
          clim.stack[[7]][[clim.var]],
          clim.stack[[8]][[clim.var]],
          clim.stack[[9]][[clim.var]],
          fun=meanIgnoringZeroes)
    return(mean.clim)
    
  }} else if(SSP=="SSP585"){
    
   AveragingClimGCMs <- function(clim.var,clim.stack){
    mean.clim <- overlay(
          clim.stack[[1]][[clim.var]],
          clim.stack[[2]][[clim.var]],
          clim.stack[[3]][[clim.var]],
          clim.stack[[4]][[clim.var]],
          clim.stack[[5]][[clim.var]],
          clim.stack[[6]][[clim.var]],
          clim.stack[[7]][[clim.var]],
          clim.stack[[8]][[clim.var]],
          fun=meanIgnoringZeroes)
    return(mean.clim)
  }   
    }

  avgbio5 <- AveragingClimGCMs(clim.var="bio5",clim.stack = stack.GCMs)
  avgbio6 <- AveragingClimGCMs(clim.var="bio6",clim.stack = stack.GCMs)
  avgbio12 <- AveragingClimGCMs(clim.var="bio12",clim.stack = stack.GCMs)
  avgbio15 <- AveragingClimGCMs(clim.var="bio15",clim.stack = stack.GCMs)


  stack.GCMs.avg <- stack(avgbio5,
                          avgbio6,
                          avgbio12,
                          avgbio15,
                          stack.GCMs[[1]][["water_top"]],
                          stack.GCMs[[1]][["depth_roots"]],
                          stack.GCMs[[1]][["TRI"]],
                          stack.GCMs[[1]][["BurnedArea"]])

  names(stack.GCMs.avg) <- c(clim.var,"water_top","depth_roots","TRI","BurnedArea")
  
  writeRaster(stack.GCMs.avg, 
              filename=paste0("data/StacksEnvVars/FutureClimate/2041to2060/",SSP,"/StackGCMsAvg.grd"),
              format="raster",
              overwrite=TRUE)
}
```


```{r ApplyAveragingGCMs}
SSPs <- c("SSP370","SSP585")

lapply(SSPs, AverageGCMs)
```




## Current climate (res 2.5 min)

We create raster stacks at 2.5 minutes resolution (resolution ~ 30.869 km2, which is the resolution of the future climates from WordlClim) with the topographic, soil, fire-related variables and **current** climatic conditions from WorldClim (i.e. from 1970 to 2000).


```{r CurrentClimateRes25min}
rast.to.crop <- stack("data/StacksEnvVars/FutureClimate/2041to2060/SSP370/Stack_BCCCSM2MR.grd")

# Climatic rasters
# ----------------
grids <- list.files("data/climate/CurrentClimate/WorldClim30sec_1970_2000/")
grids <- grids[grepl("bio_5|bio_6|bio_12|bio_15",grids)==T]
grids <- c(grids[[3]],grids[[4]],grids[[1]],grids[[2]]) # reorder
rast.clim <- raster::stack(paste0("data/climate/CurrentClimate/WorldClim30sec_1970_2000/", grids))
names(rast.clim) <- clim.var
rast.clim <- crop(rast.clim,extent(rast.to.crop)) 
rast.clim <- raster::resample(rast.clim,rast.to.crop,method="ngb")
rast.clim <- mask(rast.clim,PinpinDistri, updatevalue=NA)


# Soil rasters
# ------------
rast.soil <- stack("data/soil/STU_EU_T_TAWC_WGS84.tif","data/soil/STU_EU_DEPTH_ROOTS_WGS84.tif")
names(rast.soil) <- c("water_top","depth_roots")
rast.soil <- raster::crop(rast.soil,extent(rast.clim))   # attribute the extent of climatic raster to soil raster
rast.soil <- raster::resample(rast.soil,rast.clim)       # same number of cells between the soil and climatic rasters
rast.soil <- mask(rast.soil,PinpinDistri,updatevalue=NA) # keep only cells within maritime pine distribution 


# TRI raster
# ----------
# > The models are fitted with TRI values from raster at 90m resolution, and the 
# projections will be on raster at 1km resolution (to have the same resolution as WorldClim variables).
rast.tri <- raster("data/Topography/TRI_WGS84.tif")
rast.tri <- crop(rast.tri,extent(rast.clim))           # attribute the extent of climatic raster to TRI rasters
rast.tri <- raster::resample(rast.tri,rast.clim)       # same number of cells between the TRI and climatic rasters
names(rast.tri) <- c("TRI")
rast.tri <- mask(rast.tri,PinpinDistri,updatevalue=NA) # keep only cells within maritime pine distribution 


# Burned Area raster
# ------------------
rast.BA <- raster("data/FireData/GFED4/PreparedRasters/GFED4_199507_201412_AllCountries.grd")
rast.BA <- crop(rast.BA,extent(rast.clim))
rast.BA <- raster::resample(rast.BA,rast.clim)
names(rast.BA) <- "BurnedArea"
rast.BA <- mask(rast.BA,PinpinDistri,updatevalue=NA)


# Merge and save the rasters in a stack
# ------------------------------------
stackall <- stack(rast.clim,rast.soil,rast.tri,rast.BA)
stackall <- remove.NAs.stack(stackall) # same cells with NAs in all layers
stackall
writeRaster(stackall, filename="data/StacksEnvVars/Stack_CurrentClimate_Res25minutes.grd", format="raster",overwrite=TRUE)
```

## Adjusting missing values

```{r AdjustingNAs}
stackCurrClimRes25min <- stack("data/StacksEnvVars/Stack_CurrentClimate_Res25minutes.grd")
stackFutClimSSP370 <- stack("data/StacksEnvVars/FutureClimate/2041to2060/SSP370/StackGCMsAvg.grd")
stackFutClimSSP585 <- stack("data/StacksEnvVars/FutureClimate/2041to2060/SSP585/StackGCMsAvg.grd")

stacks <- lapply(c(stackFutClimSSP370,stackFutClimSSP585), function(x){

cellval.fut <- na.omit(as.data.frame(raster::extract(x[[1]], PinpinDistri,cellnumbers=T)))
cellval.curr <- na.omit(as.data.frame(raster::extract(stackCurrClimRes25min[[1]], PinpinDistri,cellnumbers=T)))

# 15 cells to replace by NAs in the stack of future climate
cellval.fut[cellval.fut$cell %in% setdiff(cellval.fut$cell,cellval.curr$cell),]
cell1  <- cellval.fut$cell[cellval.fut$cell %in% setdiff(cellval.fut$cell,cellval.curr$cell)]

# 4 cells to replace by NAs in stackCurrClimRes25min
cellval.curr[cellval.curr$cell %in% setdiff(cellval.curr$cell,cellval.fut$cell),]
cell2 <- cellval.curr$cell[cellval.curr$cell %in% setdiff(cellval.curr$cell,cellval.fut$cell)]


for(i in 1:8)  {x[[i]][c(cell1)] <- NA}
for(i in 1:8)  {stackCurrClimRes25min[[i]][c(cell2)] <- NA}

list(x,stackCurrClimRes25min)
})


writeRaster(stacks[[1]][[2]], filename="data/StacksEnvVars/Stack_CurrentClimate_Res25minutes.grd", 
            format="raster",overwrite=TRUE)
writeRaster(stacks[[1]][[1]], 
              filename=paste0("data/StacksEnvVars/FutureClimate/2041to2060/SSP370/StackGCMsAvg.grd"),
              format="raster",
              overwrite=TRUE)
writeRaster(stacks[[2]][[1]], 
              filename=paste0("data/StacksEnvVars/FutureClimate/2041to2060/SSP585/StackGCMsAvg.grd"),
              format="raster",
              overwrite=TRUE)
```

`r knitr::opts_chunk$set(eval = T)`


# Climate maps


For the maps, we need the hillshade for the background.

```{r LoadMapping}
background <- raster("data/maps/Hillshade/background.grd")

# Gray scale of the background:
grayscale_colors <- gray.colors(100,            # number of different color levels
                                start = 0.0,    # how black (0) to go
                                end = 1.0,      # how white (1) to go
                                gamma = 2.2,    # correction between how a digital
                                # camera sees the world and how human eyes see it
                                alpha = 0.6)
```



We load the environmental data:

  - a raster of environmental variables with **current** climate and with **resolution ~ 30.869 km2** (2.5 minutes) (5.556 x 5.556 at the equator), named: `stackCurrClim`.
  
  - a raster of environmental variables with **future** climate under scenario **SSP3-7.0** and with **resolution ~ 30.869 km2** (2.5 minutes), named: `stackFutClimSSP370`.
    
  - a raster of environmental variables with **future** climate under scenario **SSP5-8.5** and with **resolution ~ 30.869 km2** (2.5 minutes), named: `stackFutClimSSP585`.
  

```{r LoadEnvData}
stackCurrClim <- stack("data/StacksEnvVars/Stack_CurrentClimate_Res25minutes.grd")
stackFutClimSSP370 <- stack("data/StacksEnvVars/FutureClimate/2041to2060/SSP370/StackGCMsAvg.grd")
stackFutClimSSP585 <- stack("data/StacksEnvVars/FutureClimate/2041to2060/SSP585/StackGCMsAvg.grd")
```



```{r}
vars <- c("bio5","bio6","bio12","bio15")

lapply(vars, function(x){
minv <- min(c(min(values(stackCurrClim[[x]]),na.rm=T),
              min(values(stackFutClimSSP370[[x]]),na.rm=T),
              min(values(stackFutClimSSP585[[x]]),na.rm=T)))
maxv <- max(c(max(values(stackCurrClim[[x]]),na.rm=T),
              max(values(stackFutClimSSP370[[x]]),na.rm=T),
              max(values(stackFutClimSSP585[[x]]),na.rm=T)))

if(x %in% c("bio5","bio6")){
  col=colorRampPalette(c("green","blue","yellow","red","black"))
} else {
  col=colorRampPalette(c("black","red","yellow","blue","green"))
}

pdf(paste0("figs/Paper/ClimateMaps/",x,".pdf"),height=12,width=12)
layout(matrix(1:6,3,2))

# current
plot(background,col=grayscale_colors, axes=FALSE,legend=F,main="Current climate")
plot(stackCurrClim[[x]],zlim=c(minv,maxv) ,add=T,col=col(255))

# future SSP370
plot(background,col=grayscale_colors, axes=FALSE,legend=F,main="Future climate - SSP3-7.0")
plot(stackFutClimSSP370[[x]],add=T,zlim=c(minv,maxv) ,col=col(255))

# future SS
plot(background,col=grayscale_colors, axes=FALSE,legend=F,main="Future climate - SSP5-8.5")
plot(stackFutClimSSP585[[x]],add=T,zlim=c(minv,maxv) ,col=col(255))

# absolute difference btw current and SSP370
diff <- overlay(stackFutClimSSP370[[x]],
                stackCurrClim[[x]],
                fun=function(r1, r2){return(r1-r2)})

plot(background,col=grayscale_colors, axes=FALSE,legend=F,main="SSP3-7.0 - Current")
plot(diff,add=T,col=col(255))


# absolute difference btw current and SSP585
diff <- overlay(stackFutClimSSP585[[x]],
                stackCurrClim[[x]],
                fun=function(r1, r2){return(r1-r2)})

plot(background,col=grayscale_colors, axes=FALSE,legend=F,main="SSP5-8.5 - Current")
plot(diff,add=T,col=col(255))


# Percentage of change btw current and SSP585
diff <- overlay(stackFutClimSSP585[[x]],
                stackCurrClim[[x]],
                fun=function(r1, r2){r1*100/r2-100})

plot(background,col=grayscale_colors, axes=FALSE,legend=F ,
     main="Percentage change btw SSP5-8.5 and current")
plot(diff,
     add=T,
     col=col(255))
dev.off()


# Map only current climatic conditions
pdf(paste0("figs/Paper/ClimateMaps/",x,"_Current.pdf"), 
    width=18, height=18.5)
plot(background,
     col=grayscale_colors,
     axes=FALSE,legend=F)
plot(stackCurrClim[[x]],zlim=c(minv,maxv) ,add=T,col=col(255))
dev.off()

# Map only future climatic conditions
# SSP370
pdf(paste0("figs/Paper/ClimateMaps/",x,"_SSP370.pdf"), 
    width=18, height=18.5)
plot(background,
     col=grayscale_colors,
     axes=FALSE,legend=F)
plot(stackFutClimSSP370[[x]],zlim=c(minv,maxv) ,add=T,col=col(255))
dev.off()

# SSP585
pdf(paste0("figs/Paper/ClimateMaps/",x,"_SSP585.pdf"), 
    width=18, height=18.5)
plot(background,
     col=grayscale_colors,
     axes=FALSE,legend=F)
plot(stackFutClimSSP585[[x]],zlim=c(minv,maxv) ,add=T,col=col(255))
dev.off()
})
```

# References
