==> reports_Paper/8_ValidationNFI.Rmd <==
---
title: "Validating genomic offset predictions with mortality data from National Forest Inventories"
author: "Juliette Archambeau"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    df_print: kable
    highlight: textmate
    number_sections: yes
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
  pdf_document:
    toc: yes
    toc_depth: '4'
editor_options:
  chunk_output_type: console
always_allow_html: true
bibliography: references.bib 
---


<style>
pre {
  overflow-x: auto;
}
pre code {
  word-wrap: normal;
  white-space: pre;
}
</style>

<style type="text/css">
div.main-container {
  max-width: 2000px;
  margin-left: auto;
  margin-right: auto;
}
</style>


```{css, echo=FALSE}
pre {
  max-height: 300px;
  overflow-y: auto;
}

pre[class] {
  max-height: 600px;
}
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(cache=F)
options(width = 300)
library(knitr)
library(kableExtra)
library(plyr)
library(dplyr)
library(tidyverse)
library(latex2exp)
library(cowplot)
library(tidybayes)
library(rstan)
library(brms)
rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())
library(broom)
```

```{r CredibleIntervalsUsed, echo=F}
# The credible intervals used are 95% CI.
prob=0.95
```


In this document, following @capblancq2020genomic, we aimed "to validate the relationship between genetic offset and the predicted decrease in fitness (assumption c) to evaluate whether the model is indeed informative in the face of demographic complexity (i.e., the effects of other processes on spatial variance in allele frequencies, such as expansion history and gene flow) as well as the genetic architecture of climate adaptation (polygenic trait architectures, G×E interactions, nonadditive genetic variance).

For that, in this document, we want to associate the estimated genomic offset with a potential decrease in fitness in the maritime pine natural populations. We use mortality data from the French and Spanish National Forest Inventories (NFI) harmonized in @changenet2021occurrence. The French data relies on temporary plots sampled between 2005 and 2014 while the Spanish data relies on permanent plots sampled during the second (from 1986 to 1996) and third NFIs (from 1997 to 2008). A tree was recorded as dead if its death was dated at less than 5 years ago in the French NFI or if it was alive in the second inventory but dead in the third one in the Spanish NFI. 

In order to account for the different census intervals between inventories, we modeled the proportion $p_{i}$ of maritime pines that died in the plot $i$ during the census interval with the complementary log-log link and an offset on the logarithm of the census interval $\Delta_{i}$ for the plot i, as follows:

\begin{align*}
m_{i} &\sim \text{Binomial}(N_{i},p_{i})\\
\text{log}(-\text{log}(1-p_{i})) &=   \beta_{0,c} +  \beta_{C,c}C_{i} + \beta_{GO,c}GO_{i} + \text{log}(\Delta_{i}) \\
\end{align*}

with $N_{i}$ the total number of maritime pines in the plot $i$, $m_{i}$ the number of maritime pines that died during the census interval $\Delta_{i}$ in the plot $i$, $C_{i}$ the basal area of all tree species confounded in the plot $i$ (to account for the competition between trees) and $GO_{i}$ the estimated genomic offset in the plot $i$. As the French and Spanish inventories present noticeable methodological differences that may bias the estimations, we estimated country-specific coefficients: the country-specific intercepts $\beta_{0,c}$ and the country-specific slopes $\beta_{C,c}$ and $\beta_{GO,c}$.

This model will be performed for the genomic offset estimated with both the GDM and GF methods, for each set of SNPs (four sets) and each scenario of future climate (two scenarios), therefore leading to 16 models runs.

```{r CompileModel, cache=T,message=F,warning=F}
stancode = stan_model("scripts/StanModels/BinomialCloglogMortalityNFI.stan")
print(stancode)
```

# Loading and formatting NFI data

We load the NFI data harmonized in @changenet2021occurrence.

```{r LoadDataNFI}
data <- read.csv("data_DRYAD/NFIdata.csv") # 12610 obs of 337 variables
data[1:20,] %>%
  knitr::kable(digits = 3) %>% 
  kable_styling(font_size=12,bootstrap_options = c("striped","hover", "condensed"), full_width = F)
```

Variables in the dataset

  - Plot code: `plotcode`
  - Longitude and latitude of the plot: `longitude` and `latitude`.
  - Country in which the plot is: `country` (ES = Spain, FR = France).
  - The number of years between surveys in the Spanish inventory (which is equal to 5 in the French inventory as mortality is estimated in the five years before the survey date): `NbYears`.
  - The number of dead trees in the plot: `NbDead`.
  - The total number of trees in the plot: `NbTot`.
  - The basal area of all tree species in the plot (proxy of the competition among trees): `BasalArea`.




# Running the models

```{r RunModels,eval=F}
# Load the 16 genomic offset estimates in a list
# ----------------------------------------------

# load the genomic offset calculated with the GDM approach
gdm.go <- list.files("outputs_paper/GDM/ExtractedGenomicOffsetNFI/", full.names=TRUE) %>% 
  sapply(readRDS, USE.NAMES = TRUE,simplify=FALSE)
names(gdm.go) <- paste0(str_sub(names(gdm.go),46,-5),"_GDM")

# load the genomic offset calculated with the GF approach
gf.go <- list.files("outputs/GF/ExtractedGenomicOffsetNFI/", full.names=TRUE) %>% 
  sapply(readRDS, USE.NAMES = TRUE,simplify=FALSE)
names(gf.go) <- paste0(str_sub(names(gf.go),45,-5),"_GF")

# Merge the two lists
list.go <- c(gdm.go, gf.go)


# Run the 16 models
# -----------------
counter <- 0

DF.go <- sapply(list.go,function(x){
  counter <<- counter + 1 
  
  # Joining GO predictions and NFI data
  df <- data %>% 
    inner_join(x,by=c("plotcode","longitude","latitude","country","NbYears","NbDead","NbTot","BasalArea")) %>%  # 
    drop_na(GO)     # we keep only plots where we have estimations of the genomic offset
  
  stanlist <- list(N=length(df$plotcode),
                   NbDead=df$NbDead,
                   NbTot=df$NbTot,
                   logNbYears=log(df$NbYears),
                   ncountry=length(unique(df$country)),
                   country=as.numeric(df$country),
                   GO=(df$GO-mean(df$GO))/sd(df$GO),
                   C=(df$BasalArea-mean(df$BasalArea))/sd(df$BasalArea))
  
  # Running the model
  mstan <- sampling(stancode, data = stanlist, iter = 2000, chains = 4, cores = 4,init=0) 

  # Save coefficients
  broom::tidyMCMC(mstan,droppars = NULL, estimate.method = "median", 
                  ess = F, 
                  rhat = F, 
                  conf.int = T,
                  conf.level = prob) %>% 
  filter(str_detect(term, c('country')))

}, USE.NAMES = TRUE,simplify=FALSE) %>% 
  
  bind_rows(.id="GenomicOffset")

saveRDS(DF.go, file="outputs_paper/ValidationNFI/OutputsModels.rds")
```

# Visualization


```{r PaperFigure, fig.height=8,fig.width=10}
DF.go <- readRDS(file="outputs_paper/ValidationNFI/OutputsModels.rds")

model.labeller <- c("Generalised Dissimilarity Modeling", "Gradient Forests")
names(model.labeller) <- c("GDM", "GF")

graphs <- lapply(c("betaGO_country[2]","betaGO_country[1]"), function(x){

  p <-  DF.go %>% 
  filter(term==x) %>% 
  mutate(Model=str_sub(GenomicOffset,18,-1),
         SNPSets=str_sub(GenomicOffset,14,16),
         SNPSets=factor(SNPSets,levels=c("Com","Mer","Ref")),
         Scenario=str_sub(GenomicOffset,7,12))  %>%
  
  ggplot(aes(x = Scenario, y = estimate,ymin = conf.low, ymax = conf.high,colour=SNPSets,shape=SNPSets)) +
  geom_pointinterval(position = position_dodge(width = .6),point_size=2.5,size=3,show.legend = c(size = TRUE)) +
  geom_hline(yintercept = 0,color="gray") +
  facet_grid(.~Model,scales="free", space = "free",labeller = labeller(Model = model.labeller)) + 
  ylab(TeX("$\\beta_{GO}$ estimates")) + xlab("") +
  scale_colour_manual(values=c("forestgreen","chartreuse2","gold1"),labels=c("Common candidate SNPs",
            "Merged candidate SNPs",
            "Reference SNPs"),name="SNP sets") +
  scale_shape_manual(values=c(16,17,8),labels=c("Common candidate SNPs",
            "Merged candidate SNPs",
            "Reference SNPs"),name="SNP sets") +
  scale_x_discrete(labels=c("SSP370" = "SSP3-7.0", "SSP585" = "SSP5-8.5")) +
  theme_bw() +
  theme(axis.text.x = element_text(size=13),
        axis.text.y = element_text(size=16),
        axis.title = element_text(size=16),
      legend.title=element_text(size=10), 
      legend.text=element_text(size=9),
      legend.position = c(.71, .87),
      legend.background = element_rect(colour = "grey"),
      strip.text.x = element_text(size = 14),
      panel.grid.minor.x=element_blank(),
      panel.grid.major.x=element_blank()) +
    guides(color=guide_legend(ncol=3))
  
p
  
  })


graphs[[2]] <- graphs[[2]] + theme(legend.position = "none")

grid.graphs <- plot_grid(graphs[[1]],graphs[[2]],nrow=2,labels=c("A)","B)"))
ggsave(grid.graphs, file="figs/Paper/ValidationNFI/ValidationNFI.pdf",device="pdf",height=8,width=10)
ggsave(grid.graphs, file="figs/Paper/ValidationNFI/ValidationNFI.png",height=8,width=10)
grid.graphs 
```

# References
